Refactoring : Search function modularization
