+++ src/app/controller/post.controller.ts
@@ -20,46 +20,6 @@ export class PostController {
             next(error)
         }
     }
-    public async getPostByTitle(req: Request, res: Response, next: NextFunction): Promise<any> {
-        const title: string = String(req.query.title);
-        this.postService = new PostService();
-        try {
-            const exPost = await this.postService.selectPostTitle(title);
-            return res.status(200).json({
-                data: exPost
-            })
-        } catch (error) {
-            next(error)
-        }
-    }
-
-    public async getPostByEmail(req: Request, res: Response, next: NextFunction): Promise<any> {
-        const email: string = String(req.query.email);
-        console.log(email)
-        this.postService = new PostService();
-        try {
-            const exPost = await this.postService.selectPostUserEmail(email);
-            return res.status(200).json({
-                data: exPost
-            })
-        } catch (error) {
-            next(error)
-        }
-    }
-
-    public async getPostByCategory(req: Request, res: Response, next: NextFunction): Promise<any> {
-        const category: string = String(req.query.category);
-        console.log(category);
-        this.postService = new PostService();
-        try {
-            const exPost = await this.postService.selectPostCategory(category);
-            return res.status(200).json({
-                data: exPost
-            })
-        } catch (error) {
-            next(error)
-        }
-    }
 
     public async post(req: DecodedRequest, res: Response, next: NextFunction): Promise<any> {
         const userId: string = String(req.decodedId);
