+++ src/app/routes/applicationRouter.ts
@@ -5,6 +5,7 @@ import path from 'path';
 import { authRouter } from './auth.router';
 import { postRouter } from './post.router';
 import { pageRouter } from './page.router';
+import { searchRouter } from './search.router';
 
 const router: Router = express.Router();
 const swaggerSpec = YAML.load(path.join(__dirname, '../../../build/swagger.yaml'))
@@ -12,4 +13,5 @@ router.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec))
 router.use('/auth', authRouter);
 router.use('/post', postRouter);
 router.use('/page', pageRouter);
+router.use('/search', searchRouter);
 export const applicationRouter: Router = router;
